Remove management table access from progress service
