+++ app/services/targets/progress_service.rb
@@ -25,16 +25,6 @@ module Targets
       target_progress.present? ? target_progress.current_cumulative_usage_kwh : nil
     end
 
-    def management_table
-      dashboard_table = @school.latest_management_dashboard_tables.first
-      return dashboard_table.table if dashboard_table.present?
-    end
-
-    def management_data
-      dashboard_table = @school.latest_management_dashboard_tables.first
-      return dashboard_table.data if dashboard_table.present?
-    end
-
     def progress_summary
       if Targets::SchoolTargetService.targets_enabled?(@school) && target.present?
         ProgressSummary.new(
